refactor: move GLOB_IMPORTS to a separate file and update imports across components
