+++ i18n/locales/es-ES.json
@@ -758,7 +758,7 @@
       ],
       "synopsis": [
         {
-          "paragraph": "Uno, dos, tres, cuatro. ¡Uiiiiiii! Uno, dos, tres, cuatro, ¡cinco! ¡Cinco! Los hermanos Filigranes aprendieron a contar jugando con malabares. Poco a poco fueron inventando nuevas maneras de jugar… ¡y de hacerse la guitza como buenos hermanos!"
+          "paragraph": "Uno, dos, tres, cuatro. ¡Uiiiiiii! Uno, dos, tres, cuatro, ¡cinco! ¡Cinco! Los hermanos Filigranes aprendieron a contar jugando con malabares. Poco a poco fueron inventando nuevas maneras de jugar… ¡y de hacerse la travesuras como buenos hermanos!"
         },
         {
           "paragraph": "Makutu y Trinxeta nos adentrarán en su mundo particular, ofreciéndonos diferentes números de manipulación de objetos, malabares y magia cómica."
